pyproject.toml
[tool.poetry]
name = "arcgis-to-gaia"
version = "0.1.0"
description = ""
authors = ["Tony Pernicano <pernicat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
click = "^8.1.2"
python-dotenv = "^0.20.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pylint = "^2.13.5"
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
